requirements.txt
a2wsgi==1.10.0
ansi2html==1.9.1
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
connexion==3.0.5
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==3.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.3
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.4
plotly==5.18.0
pylast==5.2.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
retrying==1.3.4
rpds-py==0.16.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
spotipy==2.23.0
starlette==0.32.0.post1
tenacity==8.2.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
wikipedia==1.4.0
zipp==3.17.0
